+++ packages/access-key-management-widget/CHANGELOG.md
@@ -2,6 +2,32 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.42](https://github.com/descope/descope-js/compare/access-key-management-widget-0.1.41...access-key-management-widget-0.1.42) (2024-04-19)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.2.2`
+
+### Bug Fixes
+
+* theme mixin RELEASE ([#529](https://github.com/descope/descope-js/issues/529)) ([4e1746e](https://github.com/descope/descope-js/commit/4e1746e5db30acea0a4f8776975471d305399d8c))
+
+## [0.1.41](https://github.com/descope/descope-js/compare/access-key-management-widget-0.1.40...access-key-management-widget-0.1.41) (2024-04-18)
+
+## [0.1.40](https://github.com/descope/descope-js/compare/access-key-management-widget-0.1.39...access-key-management-widget-0.1.40) (2024-04-18)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.35`
+* `sdk-mixins` updated to version `0.2.1`
+* `sdk-component-drivers` updated to version `0.2.1`
+## [0.1.39](https://github.com/descope/descope-js/compare/access-key-management-widget-0.1.38...access-key-management-widget-0.1.39) (2024-04-18)
+
+
+### Bug Fixes
+
+* Add bound user id RELEASE ([#527](https://github.com/descope/descope-js/issues/527)) ([d245f41](https://github.com/descope/descope-js/commit/d245f413dee22f914e406338d865445adb835df2))
+
 ## [0.1.38](https://github.com/descope/descope-js/compare/access-key-management-widget-0.1.37...access-key-management-widget-0.1.38) (2024-04-15)
 
 ## [0.1.37](https://github.com/descope/descope-js/compare/access-key-management-widget-0.1.36...access-key-management-widget-0.1.37) (2024-04-15)