+++ packages/role-management-widget/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.41](https://github.com/descope/descope-js/compare/role-management-widget-0.1.40...role-management-widget-0.1.41) (2024-04-19)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.2.2`
+
+### Bug Fixes
+
+* theme mixin RELEASE ([#529](https://github.com/descope/descope-js/issues/529)) ([4e1746e](https://github.com/descope/descope-js/commit/4e1746e5db30acea0a4f8776975471d305399d8c))
+
+## [0.1.40](https://github.com/descope/descope-js/compare/role-management-widget-0.1.39...role-management-widget-0.1.40) (2024-04-18)
+
+## [0.1.39](https://github.com/descope/descope-js/compare/role-management-widget-0.1.38...role-management-widget-0.1.39) (2024-04-18)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.10.35`
+* `sdk-mixins` updated to version `0.2.1`
+* `sdk-component-drivers` updated to version `0.2.1`
 ## [0.1.38](https://github.com/descope/descope-js/compare/role-management-widget-0.1.37...role-management-widget-0.1.38) (2024-04-15)
 
 ## [0.1.37](https://github.com/descope/descope-js/compare/role-management-widget-0.1.36...role-management-widget-0.1.37) (2024-04-15)